these are the formatted XML files
added target for removing extraneous whitespaces in the programlisting tags (thanks Fabien)
	                         	SIDE EFFECTS:                             
due to the way XPath works it also removes DOCTYPEs and comments  that are not inside the root	node         (ex: comments at the beginning or end of the file)

The way the doc is structured right now the only file that contains the DOCTYPE is PA_index.xml. However PA_index.xml does not need to be formated (it is listed as excluded in the ant target).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@9969 28e8926c-6b08-0410-baaa-805c5e19b8d6