PROACTIVE-646: Fixed a NPE in RemoteObjectExposer
When a remote object is created, the remote object factory must register
the object into the registry. Since this operation is remote, it can
fail and throw an exception. This exception must be correctly handled
and throw to the user code.

This fix add a ProActiveException to some public API methods
like PAActiveObject.register(), PARemoteObject.turnActive() or
UniversalBody.registerByName(). Some IOException are now
ProActiveException

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@11937 28e8926c-6b08-0410-baaa-805c5e19b8d6