Merged PROACTIVE-648 from trunk
PROACTIVE-648: Fixed unhandled runtime exception

If the log collector is not available, then the messages are logged with
the error appender (a local FileAppender). If the log collector becomes
reachable again, a logging event is sent to notify the collector
that some logging events have been dropped. The logging events include
the timestamp of the failure, the number of dropped messages and the
original cause.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_4.1.x@11954 28e8926c-6b08-0410-baaa-805c5e19b8d6