Merged PROACTIVE-650 from trunk
PROACTIVE-650: Fixed RMISSH protocol

A RMIRemoteObjectImpl was used in the RMISSHRemoteObjectFactory due to
an inheritence issue. Since it was easy to make a such mistake, a new
AbstractRmiRemoteObjectFactory has been introduced and hold all the
business code. RMIROF and RMISSHROF use this factory and provide their
own protocolID and RMIRemoteObject. They should remain empty.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_4.1.x@11991 28e8926c-6b08-0410-baaa-805c5e19b8d6