Fixed PROACTIVE-758: UpdateCopyrightAndVersion should take in consideration tutorial-start annotations
Fixed PROACTIVE-759: Copyright updating should enable to have spaces between // and @ in the snippet annotations located before the package declaration
now, it accepts white spaces between '//' and '@'

From now on, we consider that no annotation can placed
before a copyright.
If an annotation is located before a copyright, this annotation
will be removed the next time we'll update copyrights.
Next version of toolchain will integrate a way to append the copyright
to a snippet or to a tutorial (like it was done for adding prologues to XML snippets).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@14833 28e8926c-6b08-0410-baaa-805c5e19b8d6