Added default value to PAProperty*
It is sometimes easier to define default property value in the code rather than in
the XML. If a property MUST be defined, then the code have to set a
default value if the property is not defined in the default
configuration. To avoid duplicated code (check + set), the code should define the
default value and allows overriding.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@15915 28e8926c-6b08-0410-baaa-805c5e19b8d6