Fixed DOC-57 Chapter 21. ProActive File Transfer
Actually, ProActive File Transfer had been moved to the section 15.10 (inside the Chapter 15 XML Deployment).
As a part of the File Transfer chapter is deployment-independent, this chapter has been moved outside of the XML Deployment chapter.
It is now the 16th chapter.
Nevertheless, a part of the file transfer chapter is still in the XML Deployment Chapter since it only concerns this chapter.

JIRA Links: http://bugs.activeeon.com/browse/DOC-57

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@15982 28e8926c-6b08-0410-baaa-805c5e19b8d6