Reimagined the compatibility annotations.
Groups of methods are defined at the class header. they are of the form:
@DefineGroups(
   {
   @Group(name="...", selfCompatible=true),
   ...
   }
)

And the rules concerning their compatibility can be defined right under:
@DefineRules(
   {
   @Compatible({"group1", "group2", "group3"}),
   ...
   }
)

At the level of the method, the only information needed is the belonging (or not) to a group, which is expressed with:
@MemberOf({"group1"}) -- a method can be member in more groups

This affects the implementation on the AnnotationProcessor level, where some new logic is added, also a make-over is on the way.

NOTE:
the @CompatibleWith annotation has been renamed to @Compatible
That can be applied both to methods (like in the previous methodology) or in a DefineRules section.

TODO:
Comment the annotations



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/multiactivities@18236 28e8926c-6b08-0410-baaa-805c5e19b8d6