PROACTIVE-1022 (Ibis protocol does not work)
This commit fixes the issue with ibis as transport layer.
Note that Ibis 1.4 (the one we use) can only work with Java 5 and not
Java 6 due to API changes in the java.lang.reflect package.
It should be possible to address the remaining issue by modifying the
ByteToObject and ObjectToByte classes.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@19519 28e8926c-6b08-0410-baaa-805c5e19b8d6