PROACTIVE-1051: Remove error message (IOException) printed when the PNP channel has been garbage collection
There was a race condition between channel garbage collection and
heartbeating. A tunnel can be garbage collected while the heartbeat is
sending an heartbeat leading to an error message being printed. It
should only be printed if the channel has not been closed in between.
The race condition only implied a superfluous error message.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@19890 28e8926c-6b08-0410-baaa-805c5e19b8d6