PROACTIVE-1176 GCM Deployment is broken on windows (script startn.bat is executed with invalid rmi url)
Problem:
GCM Deployment doesn't work on windows. To run new VM script 'startn.bat' is used and it is executed with invalid '-p' parameter where value of rmi url is corrupted (url starts with "rmi:\" instead of "rmi://").

Fix:
As I found that at the time when command line is constructed value of rmi url is correct, but after process is started by ProcessBuilder rmi url is getting corrupted. It seems this error was caused by the change 20532 in the class org.objectweb.proactive.extensions.gcmdeployment.GCMDeployment.Executer, in this change code starting new process was changed to use ProcessBuilder instead of Runtime.exec (ProcessBuilder's constructor expects string array, and it seems when command is passed to the ProcessBuilder as single string it treats it as name of the file to launch and modifies it, as result rmi url is changed). Code worked before with Runtime.exec because internally Runtime.exec tokenizes given command string and only after this passes the result to the ProcessBuilder, and to fix this issue I just repeated this logic in the Executer class. 


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@21441 28e8926c-6b08-0410-baaa-805c5e19b8d6