Fix PROACTIVE-1173: Interfaces declared in the controllers config file are not verified respect the implementing class.
Now it's checked on component creation. If the controller implementation does not implement the declared interface, an error message is show and the controller is ignored.

Fix PROACTIVE-1174: PAComponentRepresentativeImpl does not update the NFType when creating the NF Interfaces.

Fix PROACTIVE-1175: Manage the case when the NF Type and a Controllers Config File are both declared.

Refactored the creation of the Controller objects and the NFType of a component. Now it is possible to declared both an NFType and a Controllers Config File. The NFType takes precedence over the Controllers Config File and duplicated interfaces in the Controllers Config File are ignored. At the same time, mandatory controllers (LifeCycle, Name, Content and Binding when needed) are enforced.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@21588 28e8926c-6b08-0410-baaa-805c5e19b8d6