Upgrades to process the internal-client/internal-server roles.
PABindingBuilder: new binding type MEMBRANE_BINDING for bindings between
two interfaces that belong to the membrane. Improved management of NF
bindings, to call the appropriate version of nfBindFc, and to use the
appropriate membrane-controller, be it the owner of both components or
the client component.

PABindingBuilderItf: new parameter to specify the membrane of the owner
component.

PABindingCompiler: improved the PABindTask creation to include a task
that retrieves the membrane owner.

PABindingErrors: improved binding error messages

PATypeBindingLoader: refactored loading of bindings. Now it separate the
mappings in F interfaces, NF interfaces, and NF interfaces of
subcomponents, and descends correctly through all the tree collecting
and checking bindings. This implementation considers the possibility
that one F component and one NF component belonging to the same
component may have the same name.