PROACTIVE-1181 new ROF - AMQP - RabbitMQ 2.8.1
This commit introduces a new communication protocol based on the notion
of message queues. The implementation supports the specification
AMQP-0.9.1 as we use rabbitmq-client in version 2.8.1.
So we can connect to any AMQP-0.9.1 compliant server.

We do not provide an AMQP server in the distribution, only the ability
to connect to a server. Server installation or configuration are not described
here.

This implementation has to be considered as a prototype and should not
be used in production so far.