PROACTIVE-1243 Add client eviction mechanism
Problem:

The map referenced by the clientMap field of RouterImpl is never
cleaned. As a result, after a certain number of connected and
disconnected clients, the router dies with an OOME.

Periodic connections/disconnections of clients are common in scenarios
when nodes are started/stopped using a cron-like policy in the RM or
agents. Another scenario are the dynamic node sources with nodes
started/stopped on demand.

In the above scenarios, the reserved agent IDs are usually not used,
which leads to a new entry allocated in the clientMap for every
started node. As a result, the size of the map grows linearly with
time.

With the heap size of 500 MB, the heap becomes full when the agent IDs
reach ~200000.

Fix:

Introduce a client eviction mechanism. A user can use the command line
option -e to specify the timeout in ms after which the disconnected
clients will be removed from the clientMap. The eviction is disabled
if the timeout is -1 (default). Reserved clients are never evicted.

Since the eviction is done without synchronization, it is possible for
a just-reconnected client to be evicted. Such a client will fail to
communicate properly and will get a tunnel failure. I think it is an
acceptable alternative to just failing to reconnect.

Change-Id: Ifaf5fabe2d7d4e605d5d1a89fb11b4d8892f53f6