PROACTIVE-1258 pamr: fix the heartbeat mechanism
Problem:

Heartbeat mechanism was unable to detect the death of a
message reader thread in an agent. As a result, any call to a remote
object managed by such agent would hang indefinitely. Message reader
thread is particularly prone to death under low limit on the number of
user threads (ulimit -u) because it does submissions to a thread pool,
and thus, thread creations, which can fail with

    java.lang.OutOfMemoryError: unable to create new native thread

Fix:

Instead of sending agent->router heartbeats based on a timer, send
them as replies to router->agent heartbeats. Thus, the connectivity
between the agent and the router is checked in full. The router is
then able to detect agent failure and take action.

Change-Id: I1f514b04f06566f54f115ee7529cd9458b2cedb5