PROACTIVE-1324 - Improve request interruption capabilities
If we want to interrupt an active object serving a request, we stumble on the following problems :
Currently it is very complicated to get a handle on an active object activity thread
the InterruptedException are catched inside the body serving mechanism and
are not forwarded. Thus, someone writing a runActivity cannot decide how
to react to an interruption. This add the interruptService method to allow
local or remote interruption of an AO activity. The call only interrupts the
current request served, it does not terminate the Active Object

Added a functional test

Change-Id: I467586e63a3792d52203155dac7fa6638831ba0c