Fix PNP/Dataspace stacktrace printing on Scheduler startup
When using multi protocols a stack trace was printed from SweetCountDownLatch.
It is when using dataspaces and multiprotocols, the URLs are tested in parallel
for each protocol and only one is required.
Once was found the other attempts were interrupted, causing the stacktrace.

Stacktrace printing has been replaced by logging.

Change-Id: Id71710182891b5bab614e9188620514b44a241f7