SCHEDULING-2165 SCHEDULING-2166 Runasme on Mac OS X
The Bash scripts were not working on Mac due
 - mktemp has a different syntax
 - uname has a different output/syntax

suer program was not working on Mac due to
 - differences in su command (stream used)
 - su location
 - include header
A version has been created, specific to Mac

Also this runAsMe/suer feature relies on temporary files, on Mac the default
tmp folder is not accessible to everyone. Therefore the $TMPDIR environment
variable and Java system property java.io.tmpdir should be set to /tmp
when running a node with fork method = pwd.

Change-Id: Ifd6d09f9dee2aa6094f88a644deeb44a3241584f