SCHEDULING-2253 serialVersionUID plugin broken after upgrade to Java 7
The Gradle serialver plugin we use did not work anymore after the upgrade
to Java 7 so we changed it, it now relies on .classes file, therefore
needs to be executed after stub are generated.

Javassist was upgraded to be aligned with the one used by the plugin and
to avoid conflicts.