ProActiveRuntime#getActiveObjects returned wrong list of active objects
Because the list of bodies was traversed while deleting elements, an incompete
list of active objects could be returned.

For instance if a body at pos=3 was removed, then the list is shifted, and
for the next iteration, element that was at pos=4 could be missed.

This was preventing node cleaning to be performed, thus breaking node cleaning
in the Scheduler as well as observed with tests such as TestJobRemoved.